Restructuring. Stubbing tests. Adding makefile.
