+++ python/Makefile
@@ -1,0 +1,7 @@
+.PHONY: test
+
+test:
+	nosetests
+
+clean:
+	find
