+++ python/requirements.txt
@@ -1,0 +1,3 @@
+argparse==1.2.1
+nose==1.3.4
+wsgiref==0.1.2
