+++ python/sudoku_solver/cell.py
@@ -1,0 +1,8 @@
+
+
+class Cell(object):
+    def __init__(self, value):
+        self.value = value
+
+    def __str__(self):
+        return str(self.value)
