+++ python/tests/test_group.py
@@ -1,0 +1,5 @@
+from unittest import TestCase
+
+class GroupTestCase(TestCase):
+    def test_group(self):
+        pass
