+++ python/tests/test_puzzle.py
@@ -1,0 +1,5 @@
+from unittest import TestCase
+
+class PuzzleTestCase(TestCase):
+    def test_puzzle(self):
+        pass
